Rename the package scripts according to their designation to sensible names
